tatiana_mantrova/test-requirements.txt
pytest >= 8.4.1
pytest-cov >= 2.8.1
tox >= 3.9.0
flake8 >= 4.0.0
types-python-dateutil >= 2.8.19.14
mypy >= 1.5
python-dotenv >= 1.1.1
